Add `jump()` function
(Compare with the changes in the 'navigation' PR in the 'mint-runtime' repo.)

User can now choose between `navigate()`, which never jumps, and `jump()` which does, if the URL is a defined route.

The parameter lists and descriptions of the related functions `navigate()` and `setUrl()` are also updated. Renamed `triggerHashJump()` to `triggerJump()`, to fit more to the name of `jump()`.